Extract deterministic_fallback_plan.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/deterministic_fallback_plan.rs`
to host the deterministic fallback-plan generator used by the
planning-model timeout recovery path. Small follow-on extraction in
the meta-issue #680 reduction effort.

Migrated to `deterministic_fallback_plan.rs` (`pub(super)`):

* `extract_requested_port(task) -> Option<String>` (first 2-5 digit run)
* `fallback_plan_request_label(task) -> String`
* `fallback_plan_platform_label(task) -> &'static str`
* `deterministic_timeout_fallback_plan(task, &TaskProfile, work_root)
    -> String` (renders the markdown plan with Goal / Constraints /
    First Action / Verification sections, parameterised on detected
    port + framework + worktree name + task profile)

`loop_run.rs` adjustments:

* Registered `mod deterministic_fallback_plan;` with a parent-#680
  doc comment.

`turn.rs` adjustments:

* Removed the 4 fn definitions (originally at lines 604-667).
* Pruned `deterministic_timeout_fallback_plan` from existing
  test-mod `use super::{...}` import block.
* Pruned now-unused `TaskProfile` from the existing
  `use crate::modes::plan_act::{...}` import block (the symbol is
  still used by `deterministic_fallback_plan` and in the test mod
  via its own `use crate::modes::plan_act::{...}`).
* Test mod gained a `use super::super::deterministic_fallback_plan::deterministic_timeout_fallback_plan;`
  inside the single test that exercises it.

`scaffold_pipeline.rs` adjustments:

* Updated `deterministic_timeout_fallback_plan` import from
  `super::turn::{...}` group → standalone
  `use super::deterministic_fallback_plan::deterministic_timeout_fallback_plan;`.

LOC delta:

* `turn.rs`: 28,054 → 27,987 (-67 LOC)
* `deterministic_fallback_plan.rs`: 0 → 88 (new)
* Cumulative `turn.rs`: 39,489 → 27,987 (-11,502 LOC, -29.1%)

`pub(super)` limited / no facade re-export (DR3-001).

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
